#256 reorganise components
