+++ src/components/ContentfulModule/EventsModule/EventCard/index.js
