+++ src/components/ContentfulModule/EventsModule/EventCardList/index.js
