+++ src/components/ContentfulModule/EventsModule/EventsModule.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { EventsModulePropTypes } from "./EventsModulePropTypes";
-import EventCardList from "../../EventCardList";
-import EventCard from "../../EventCard";
+import EventCardList from "./EventCardList";
+import EventCard from "./EventCard";
 import DetailsWrapper from "../../DetailsWrapper";
 import { getTranslatedText } from "../../../utils/getTranslatedText";
