+++ src/components/Organisation/Card/index.js
@@ -1,0 +1,4 @@
+import Card from "./Card";
+import CardSection from "./CardSection";
+
+export { Card, CardSection };
