+++ src/components/Organisation/index.js
@@ -1,4 +1,6 @@
 import CardListSection from "./CardListSection";
+import CardList from "./CardList";
+import { Card, CardSection } from "./Card";
 import { OrganisationPropTypes } from "./OrganisationPropTypes";
 
-export { CardListSection, OrganisationPropTypes };
+export { CardListSection, OrganisationPropTypes, CardList, Card, CardSection };
